Extract FeedCacheSpy to shared scopre to remove duplication
